fix: restore exact original UI code in decomposed files, revert EPUB split
- ReaderModeComponents: ReadModeSection/DoublePageConfigSection restored verbatim (34dp stacked icons, no ripple, reader_mode_hint, indented sub-options)
- RelatedSection: restore maskClip carousel masking
- Pill: restore text ellipsis; LoadingHero: remove stray spacer
- Revert EpubReaderFragment decomposition to avoid merge conflicts
