+++ app/src/main/kotlin/org/koitharu/kotatsu/details/ui/DetailsCommonComponents.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import org.koitharu.kotatsu.R
 
@@ -109,6 +110,7 @@ internal fun Pill(
 				fontWeight = FontWeight.Medium,
 				color = content,
 				maxLines = 1,
+				overflow = TextOverflow.Ellipsis,
 			)
 		}
 	}
@@ -116,7 +118,6 @@ internal fun Pill(
 
 @Composable
 internal fun LoadingHero() {
-	Spacer(Modifier.height(8.dp))
 	Box(
 		modifier = Modifier
 			.fillMaxWidth()
