+++ app/src/main/kotlin/org/koitharu/kotatsu/reader/ui/config/ReaderModeComponents.kt
@@ -1,5 +1,6 @@
 package org.koitharu.kotatsu.reader.ui.config
 
+import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
@@ -11,6 +12,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -25,6 +27,7 @@ import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.BiasAlignment
 import androidx.compose.ui.Modifier
@@ -33,221 +36,249 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import org.koitharu.kotatsu.R
 import org.koitharu.kotatsu.core.prefs.ReaderMode
+import kotlin.math.roundToInt
 
 @Composable
 internal fun ReadModeSection(
-	selectedMode: ReaderMode,
-	onModeSelected: (ReaderMode) -> Unit,
+    selectedMode: ReaderMode,
+    onModeSelected: (ReaderMode) -> Unit,
 ) {
-	val modes = listOf(
-		ReaderMode.STANDARD to (R.string.standard to R.drawable.ic_reader_ltr),
-		ReaderMode.REVERSED to (R.string.r_to_l to R.drawable.ic_reader_rtl),
-		ReaderMode.VERTICAL to (R.string.vertical to R.drawable.ic_reader_vertical),
-		ReaderMode.WEBTOON to (R.string.webtoon to R.drawable.ic_reader_webtoon),
-	)
-	val selectedIndex = modes.indexOfFirst { it.first == selectedMode }.coerceAtLeast(0)
+    val modes = listOf(
+        ReaderMode.STANDARD to (R.string.standard to R.drawable.ic_reader_ltr),
+        ReaderMode.REVERSED to (R.string.r_to_l to R.drawable.ic_reader_rtl),
+        ReaderMode.VERTICAL to (R.string.vertical to R.drawable.ic_reader_vertical),
+        ReaderMode.WEBTOON to (R.string.webtoon to R.drawable.ic_reader_webtoon),
+    )
+    val selectedIndex = modes.indexOfFirst { it.first == selectedMode }.coerceAtLeast(0)
 
-	Column(
-		modifier = Modifier
-			.fillMaxWidth()
-			.padding(horizontal = 16.dp),
-		verticalArrangement = Arrangement.spacedBy(8.dp),
-	) {
-		Surface(
-			shape = RoundedCornerShape(24.dp),
-			color = MaterialTheme.colorScheme.surfaceContainerHigh,
-			modifier = Modifier.fillMaxWidth(),
-		) {
-			Box(
-				modifier = Modifier
-					.fillMaxWidth()
-					.height(112.dp)
-					.padding(8.dp),
-			) {
-				val targetBias = when (selectedIndex) {
-					0 -> -1f
-					1 -> -1f / 3f
-					2 -> 1f / 3f
-					3 -> 1f
-					else -> -1f
-				}
-				val animatedBias by animateFloatAsState(
-					targetValue = targetBias,
-					animationSpec = tween(durationMillis = 300, easing = FastOutSlowInEasing),
-					label = "mode_highlighter",
-				)
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        Surface(
+            shape = RoundedCornerShape(24.dp),
+            color = MaterialTheme.colorScheme.surfaceContainerHigh,
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(112.dp)
+                    .padding(8.dp),
+            ) {
+                val targetBias = when (selectedIndex) {
+                    0 -> -1f
+                    1 -> -1f / 3f
+                    2 -> 1f / 3f
+                    3 -> 1f
+                    else -> -1f
+                }
+                val animatedBias by animateFloatAsState(
+                    targetValue = targetBias,
+                    animationSpec = tween(durationMillis = 300, easing = FastOutSlowInEasing),
+                    label = "mode_highlighter",
+                )
 
-				Box(
-					modifier = Modifier
-						.fillMaxHeight()
-						.fillMaxWidth(0.25f)
-						.align(BiasAlignment(horizontalBias = animatedBias, verticalBias = 0f))
-						.clip(RoundedCornerShape(22.dp))
-						.background(MaterialTheme.colorScheme.primary),
-				)
+                Box(
+                    modifier = Modifier
+                        .fillMaxHeight()
+                        .fillMaxWidth(0.25f)
+                        .align(BiasAlignment(horizontalBias = animatedBias, verticalBias = 0f))
+                        .clip(RoundedCornerShape(22.dp))
+                        .background(MaterialTheme.colorScheme.primary),
+                )
 
-				Row(
-					modifier = Modifier.fillMaxWidth(),
-					verticalAlignment = Alignment.CenterVertically,
-				) {
-					modes.forEach { (mode, pairRes) ->
-						val (labelRes, iconRes) = pairRes
-						val isSelected = selectedMode == mode
-						val contentColor by animateFloatAsState(
-							targetValue = if (isSelected) 1f else 0.6f,
-							animationSpec = tween(durationMillis = 300),
-							label = "mode_content_alpha",
-						)
+                Row(
+                    modifier = Modifier.fillMaxSize(),
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    modes.forEachIndexed { index, (mode, pair) ->
+                        val (labelRes, iconRes) = pair
+                        val isSelected = selectedMode == mode
+                        val contentColor by animateColorAsState(
+                            targetValue = if (isSelected) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurfaceVariant,
+                            label = "segment_fg_$index",
+                        )
 
-						Box(
-							modifier = Modifier
-								.weight(1f)
-								.fillMaxHeight()
-								.clickable { onModeSelected(mode) },
-							contentAlignment = Alignment.Center,
-						) {
-							Row(
-								verticalAlignment = Alignment.CenterVertically,
-								horizontalArrangement = Arrangement.Center,
-							) {
-								Icon(
-									painter = painterResource(iconRes),
-									contentDescription = null,
-									tint = (if (isSelected) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurface).copy(alpha = contentColor),
-									modifier = Modifier.size(18.dp),
-								)
-								Spacer(modifier = Modifier.width(6.dp))
-								Text(
-									text = stringResource(labelRes),
-									style = MaterialTheme.typography.titleSmall,
-									fontWeight = FontWeight.Bold,
-									color = (if (isSelected) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurface).copy(alpha = contentColor),
-								)
-							}
-						}
-					}
-				}
-			}
-		}
-	}
+                        Box(
+                            modifier = Modifier
+                                .weight(1f)
+                                .fillMaxHeight()
+                                .clickable(
+                                    interactionSource = remember { androidx.compose.foundation.interaction.MutableInteractionSource() },
+                                    indication = null,
+                                ) { onModeSelected(mode) },
+                            contentAlignment = Alignment.Center,
+                        ) {
+                            Column(
+                                horizontalAlignment = Alignment.CenterHorizontally,
+                                verticalArrangement = Arrangement.Center,
+                            ) {
+                                Icon(
+                                    painter = painterResource(iconRes),
+                                    contentDescription = null,
+                                    modifier = Modifier.size(34.dp),
+                                    tint = contentColor,
+                                )
+                                Spacer(modifier = Modifier.height(8.dp))
+                                Text(
+                                    text = stringResource(labelRes),
+                                    style = MaterialTheme.typography.titleSmall,
+                                    fontWeight = FontWeight.Medium,
+                                    maxLines = 1,
+                                    overflow = TextOverflow.Ellipsis,
+                                    color = contentColor,
+                                )
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        Text(
+            text = stringResource(R.string.reader_mode_hint),
+            style = MaterialTheme.typography.bodySmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+            modifier = Modifier.padding(horizontal = 8.dp),
+        )
+    }
 }
 
 @Composable
 internal fun DoublePageConfigSection(
-	isModeStandardOrReversed: Boolean,
-	isDoubleOnLandscape: Boolean,
-	onDoubleOnLandscapeChange: (Boolean) -> Unit,
-	isDoubleOnFoldable: Boolean,
-	onDoubleOnFoldableChange: (Boolean) -> Unit,
-	sensitivity: Float,
-	onSensitivityChange: (Float) -> Unit,
+    isModeStandardOrReversed: Boolean,
+    isDoubleOnLandscape: Boolean,
+    onDoubleOnLandscapeChange: (Boolean) -> Unit,
+    isDoubleOnFoldable: Boolean,
+    onDoubleOnFoldableChange: (Boolean) -> Unit,
+    sensitivity: Float,
+    onSensitivityChange: (Float) -> Unit,
 ) {
-	val sectionAlpha = if (isModeStandardOrReversed) 1f else 0.38f
+    val sectionAlpha = if (isModeStandardOrReversed) 1f else 0.38f
 
-	Surface(
-		shape = RoundedCornerShape(24.dp),
-		color = MaterialTheme.colorScheme.surfaceContainerHigh,
-		border = androidx.compose.foundation.BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.55f)),
-		shadowElevation = 1.dp,
-		modifier = Modifier
-			.fillMaxWidth()
-			.padding(horizontal = 16.dp),
-	) {
-		Column(
-			modifier = Modifier
-				.fillMaxWidth()
-				.alpha(sectionAlpha)
-				.padding(vertical = 12.dp),
-			verticalArrangement = Arrangement.spacedBy(8.dp),
-		) {
-			Row(
-				modifier = Modifier
-					.fillMaxWidth()
-					.clickable(enabled = isModeStandardOrReversed) { onDoubleOnLandscapeChange(!isDoubleOnLandscape) }
-					.padding(horizontal = 20.dp, vertical = 10.dp),
-				verticalAlignment = Alignment.CenterVertically,
-			) {
-				Icon(
-					painter = painterResource(R.drawable.ic_split_horizontal),
-					contentDescription = null,
-					tint = MaterialTheme.colorScheme.primary,
-					modifier = Modifier.size(22.dp),
-				)
-				Spacer(modifier = Modifier.width(16.dp))
-				Text(
-					text = stringResource(R.string.use_two_pages_landscape),
-					style = MaterialTheme.typography.titleMedium,
-					color = MaterialTheme.colorScheme.onSurface,
-					modifier = Modifier.weight(1f),
-				)
-				Switch(
-					checked = isDoubleOnLandscape,
-					onCheckedChange = onDoubleOnLandscapeChange,
-					enabled = isModeStandardOrReversed,
-				)
-			}
+    Surface(
+        shape = RoundedCornerShape(24.dp),
+        color = MaterialTheme.colorScheme.surfaceContainerHigh,
+        border = androidx.compose.foundation.BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.55f)),
+        shadowElevation = 1.dp,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp),
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .alpha(sectionAlpha)
+                .padding(vertical = 12.dp),
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+        ) {
+            // Toggle 1: Use two pages in landscape
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable(enabled = isModeStandardOrReversed) { onDoubleOnLandscapeChange(!isDoubleOnLandscape) }
+                    .padding(horizontal = 20.dp, vertical = 10.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Icon(
+                    painter = painterResource(R.drawable.ic_split_horizontal),
+                    contentDescription = null,
+                    tint = MaterialTheme.colorScheme.primary,
+                    modifier = Modifier.size(22.dp),
+                )
+                Spacer(modifier = Modifier.width(16.dp))
+                Text(
+                    text = stringResource(R.string.use_two_pages_landscape),
+                    style = MaterialTheme.typography.titleMedium,
+                    color = MaterialTheme.colorScheme.onSurface,
+                    modifier = Modifier.weight(1f),
+                )
+                Switch(
+                    checked = isDoubleOnLandscape,
+                    onCheckedChange = onDoubleOnLandscapeChange,
+                    enabled = isModeStandardOrReversed,
+                )
+            }
 
-			val subOptionsAlpha = if (isModeStandardOrReversed && isDoubleOnLandscape) 1f else 0.38f
+            // Sub-options
+            val subOptionsAlpha = if (isModeStandardOrReversed && isDoubleOnLandscape) 1f else 0.38f
 
-			Column(
-				modifier = Modifier
-					.fillMaxWidth()
-					.alpha(subOptionsAlpha),
-				verticalArrangement = Arrangement.spacedBy(8.dp),
-			) {
-				Row(
-					modifier = Modifier
-						.fillMaxWidth()
-						.clickable(enabled = isModeStandardOrReversed && isDoubleOnLandscape) { onDoubleOnFoldableChange(!isDoubleOnFoldable) }
-						.padding(horizontal = 20.dp, vertical = 10.dp),
-					verticalAlignment = Alignment.CenterVertically,
-				) {
-					Spacer(modifier = Modifier.width(38.dp))
-					Text(
-						text = stringResource(R.string.auto_double_foldable),
-						style = MaterialTheme.typography.titleMedium,
-						color = MaterialTheme.colorScheme.onSurface,
-						modifier = Modifier.weight(1f),
-					)
-					Switch(
-						checked = isDoubleOnFoldable,
-						onCheckedChange = onDoubleOnFoldableChange,
-						enabled = isModeStandardOrReversed && isDoubleOnLandscape,
-					)
-				}
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .alpha(subOptionsAlpha),
+                verticalArrangement = Arrangement.spacedBy(8.dp),
+            ) {
+                // Toggle 2: Auto double on foldable
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .clickable(enabled = isModeStandardOrReversed && isDoubleOnLandscape) { onDoubleOnFoldableChange(!isDoubleOnFoldable) }
+                        .padding(horizontal = 20.dp, vertical = 10.dp),
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    Spacer(modifier = Modifier.width(38.dp)) // indentation for sub-option
+                    Text(
+                        text = stringResource(R.string.auto_double_foldable),
+                        style = MaterialTheme.typography.titleMedium,
+                        color = MaterialTheme.colorScheme.onSurface,
+                        modifier = Modifier.weight(1f),
+                    )
+                    Switch(
+                        checked = isDoubleOnFoldable,
+                        onCheckedChange = onDoubleOnFoldableChange,
+                        enabled = isModeStandardOrReversed && isDoubleOnLandscape,
+                    )
+                }
 
-				Column(
-					modifier = Modifier
-						.fillMaxWidth()
-						.padding(horizontal = 20.dp, vertical = 8.dp),
-				) {
-					Row(
-						modifier = Modifier.fillMaxWidth(),
-						horizontalArrangement = Arrangement.SpaceBetween,
-						verticalAlignment = Alignment.CenterVertically,
-					) {
-						Text(
-							text = stringResource(R.string.two_page_scroll_sensitivity),
-							style = MaterialTheme.typography.titleMedium,
-							color = MaterialTheme.colorScheme.onSurface,
-						)
-						Text(
-							text = "${sensitivity.toInt()}%",
-							style = MaterialTheme.typography.bodyMedium,
-							color = MaterialTheme.colorScheme.onSurfaceVariant,
-						)
-					}
-					Slider(
-						value = sensitivity,
-						onValueChange = onSensitivityChange,
-						valueRange = 0f..100f,
-						enabled = isModeStandardOrReversed && isDoubleOnLandscape,
-					)
-				}
-			}
-		}
-	}
+                // Slider: Scroll sensitivity
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 20.dp, vertical = 8.dp),
+                ) {
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.SpaceBetween,
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        Row(verticalAlignment = Alignment.CenterVertically) {
+                            Spacer(modifier = Modifier.width(38.dp))
+                            Text(
+                                text = stringResource(R.string.two_page_scroll_sensitivity),
+                                style = MaterialTheme.typography.titleMedium,
+                                color = MaterialTheme.colorScheme.onSurface,
+                            )
+                        }
+                        Text(
+                            text = "${sensitivity.roundToInt()}%",
+                            style = MaterialTheme.typography.titleMedium,
+                            fontWeight = FontWeight.Bold,
+                            color = if (isModeStandardOrReversed && isDoubleOnLandscape) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.38f),
+                        )
+                    }
+
+                    Spacer(modifier = Modifier.height(4.dp))
+
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Spacer(modifier = Modifier.width(38.dp))
+                        Slider(
+                            value = sensitivity,
+                            onValueChange = onSensitivityChange,
+                            valueRange = 0f..100f,
+                            enabled = isModeStandardOrReversed && isDoubleOnLandscape,
+                            modifier = Modifier.weight(1f),
+                        )
+                    }
+                }
+            }
+        }
+    }
 }
